+++ tools/onnx-graphsurgeon/onnx_graphsurgeon/ir/graph.py
@@ -99,6 +99,7 @@ class Graph(object):
         doc_string=None,
         opset=None,
         import_domains=None,
+        ir_version=None,
     ):
         """
         Args:
@@ -107,6 +108,7 @@ class Graph(object):
             outputs (Sequence[Tensor]): A list of graph output Tensors.
             name (str): The name of the graph. Defaults to "onnx_graphsurgeon_graph".
             doc_string (str): A doc_string for the graph. Defaults to "".
+            opset (int): The ONNX opset to use when exporting this graph.
         """
         self.nodes = misc.default_value(nodes, [])
         self.inputs = list(misc.default_value(inputs, []))
@@ -117,7 +119,8 @@ class Graph(object):
 
         self.doc_string = misc.default_value(doc_string, "")
         self.opset = misc.default_value(opset, Graph.DEFAULT_OPSET)
-        self.import_domains = misc.default_value(import_domains, None)
+        self.import_domains = import_domains
+        self.ir_version = ir_version
         # Printing graphs can be very expensive
         G_LOGGER.ultra_verbose(lambda: "Created Graph: {:}".format(self))
         # For layer() function
@@ -481,6 +484,81 @@ class Graph(object):
         PARTITIONING_MODES = [None, "basic", "recursive"]
         if partitioning not in PARTITIONING_MODES:
             G_LOGGER.critical("Argument for parameter 'partitioning' must be one of: {:}".format(PARTITIONING_MODES))
+
+        # First perform shape tensor cast elision on the graph prior to other constant folding
+        # Search for Cast(s) (from int -> float) -> intermediate operator (with float constants) -> Cast(s) (back to int)
+        # This pattern is problematic for TensorRT since these operations may be performed on Shape Tensors, which
+        # are not allowed to be floating point type. Attempt to fold the pattern here
+        VALID_CAST_ELISION_OPS = ["Add", "Sub", "Mul", "Div", "Max", "Min", "Equal", "Greater", "Less", "Concat"]
+
+        def run_cast_elision(node):
+            import onnx
+
+            if node.op not in VALID_CAST_ELISION_OPS:
+                return
+
+            # Get list of input nodes
+            inp_casts = [
+                inp_node
+                for inp_tensor in node.inputs
+                for inp_node in inp_tensor.inputs
+                if inp_node.op == "Cast" and inp_node.attrs["to"] == 1
+            ]
+
+            # No cast nodes found, return early
+            if not inp_casts:
+                return
+
+            # Ensure that all input cast nodes are casting from the same type
+            final_type = None
+            for inp in inp_casts:
+                curr_type = onnx.mapping.NP_TYPE_TO_TENSOR_TYPE[inp.inputs[0].dtype]
+                final_type = final_type or curr_type
+                if final_type != curr_type:
+                    return
+
+            # Check validity and get list of output nodes
+            out_casts = []
+
+            for out_tensor in node.outputs:
+                for out_node in out_tensor.outputs:
+                    if out_node.op != "Cast" or out_node.attrs["to"] not in [6, 7]:
+                        # Can exit early if any of the output nodes are not valid casts
+                        return
+                    out_casts.append(out_node)
+                    # Check that all final cast types are the same.
+                    curr_type = out_node.attrs["to"]
+                    if final_type != curr_type:
+                        return
+
+            # If all checks passed - update constant values.
+            for inp in node.inputs:
+                if isinstance(inp, Constant):
+                    inp.values = inp.values.astype(onnx.mapping.TENSOR_TYPE_TO_NP_TYPE[final_type])
+
+            # "Remove" casts nodes by changing I/O node operators to Identity. Update corresponding tensor dtypes as well
+            def replace_with_identity(cast_node, change_dtype):
+                cast_node.op = "Identity"
+                cast_node.attrs = {}
+                getattr(cast_node, change_dtype)[0].dtype = onnx.mapping.TENSOR_TYPE_TO_NP_TYPE[final_type]
+                G_LOGGER.debug("Cast node {:} elided".format(cast_node.name))
+
+            for inp in inp_casts:
+                replace_with_identity(inp, change_dtype="outputs")
+
+            for out in out_casts:
+                replace_with_identity(out, change_dtype="inputs")
+
+        # Perform shape tensor cast elision:
+        if fold_shapes:
+            G_LOGGER.debug("Performing shape tensor cast elision in {:}".format(self.name))
+            try:
+                for node in self.nodes:
+                    run_cast_elision(node)
+            except Exception as err:
+                if not error_ok:
+                    raise err
+                G_LOGGER.warning("'{:}' routine failed with: {:}".format("Shape tensor cast elision", err))
 
         G_LOGGER.debug("Folding constants in {:}".format(self.name))